refactored KBtn example
